pyproject.toml
[tool.poetry]
name = "crawjud-bots"
version = "8.0.0"
description = "CrawJUD API para execução de bots"
authors = ["Robotz213 <nicholasazs213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask-socketio = "^5.4.1"
flask = "^3.0.3"
flask-mysqldb = "^2.0.0"
flask-mail = "^0.10.0"
selenium = "^4.26.1"
webdriver-manager = "^4.0.2"
google-cloud-storage = "^2.18.2"
google-auth = "^2.36.0"
openpyxl = "^3.1.5"
pypdf2 = "^3.0.1"
tqdm = "^4.67.0"
clear = "^2.0.0"
psutil = "^6.1.0"
bcrypt = "^4.2.0"
python-socketio = "^5.11.4"
flask-sqlalchemy = "^3.1.1"
pywinauto = "^0.6.8"
pillow = "^11.0.0"
python-docx = "^1.1.2"
pymupdf = "^1.24.13"
pandas = "^2.2.3"
flask-talisman = "^1.1.0"
pyyaml = "^6.0.2"
termcolor = "^2.5.0"
psycopg2 = "^2.9.10"
rich = "^13.9.4"
deep-translator = "^1.11.4"
chrome-version = "^0.3.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
inquirer = "^3.4.0"
findmyip = "^1.2.0"
aiofiles = "^24.1.0"
watchdog = "^6.0.0"
gevent = "^24.10.3"
gevent-websocket = "^0.10.1"
debugpy = "^1.8.7"
gitpython = "^3.1.43"
pygithub = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
